disassemble sbom file
