+++ api/app/sbom/parser/debug_info_outputer.py
@@ -1,0 +1,3 @@
+def error_message(*args, **kwargs):
+    # print(*args, **{k: v for (k, v) in kwargs if k != "file"}, file=sys.stderr)
+    pass
