+++ api/app/sbom/parser/sbom_info.py
@@ -1,0 +1,10 @@
+from typing import (
+    NamedTuple,
+)
+
+
+class SBOMInfo(NamedTuple):
+    spec_name: str
+    spec_version: str
+    tool_name: str
+    tool_version: str | None
